delete customer and search customer added

+++ src/main/webapp/WEB-INF/view/list-customers.jsp
@@ -1,4 +1,5 @@
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
+<%@ taglib prefix="form" uri="http://www.springframework.org/tags/form" %>
 <%--
   Created by IntelliJ IDEA.
   User: ahmed
@@ -20,7 +21,25 @@
 </head>
 <body>
 <h1 class="test">List of Customer of Our Company</h1>
-<a href="${pageContext.request.contextPath}/customer/showForm" class="btn btn-primary m-4" >Add Customer</a>
+<div class="row">
+    <div class="col-12 ml-3">
+        <form:form action="search" method="GET">
+            <div class="row">
+                <div class="col-3">
+                    <input type="text" class="form-control" name="theSearchName">
+                </div>
+                <div class="col-3">
+                    <button type="submit" class="btn btn-dark">Search</button>
+                </div>
+            </div>
+        </form:form>
+    </div>
+    <div class="col-3">
+        <a href="${pageContext.request.contextPath}/customer/showForm" class="btn btn-primary m-4" >Add Customer</a>
+    </div>
+
+</div>
+
 <div class="wrapper">
     <div class="header">
         <h1>CRM Table</h1>
@@ -42,15 +61,15 @@
                 <c:url var="updateLink" value="/customer/showFormForUpdate">
                     <c:param name="customerID" value="${tempCustomer.id}"/>
                 </c:url>
-                <c:url var="deleteLink" value="/customer/showFormForUpdate">
-                    <c:param name="customerId" value="${tempCustomer.id}"/>
+                <c:url var="deleteLink" value="/customer/delete">
+                    <c:param name="customerID" value="${tempCustomer.id}"/>
                 </c:url>
                 <tr>
                     <td>${tempCustomer.firstName}</td>
                     <td>${tempCustomer.lastName}</td>
                     <td>${tempCustomer.email}</td>
                     <td><a href="${updateLink}" class="btn btn-warning">Update</a></td>
-                    <td><a href="${deleteLink}" class="btn btn-danger" >Delete</a></td>
+                    <td><a href="${deleteLink}" onclick="if(!(confirm('Are you sure to delete this customer ?'))) return false" class="btn btn-danger" >Delete</a></td>
                 </tr>
             </c:forEach>
         </table>
